pyproject.toml
[project]
name = "partij-programma-wijzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "flake8-pyproject>=1.2.3",
    "google-genai>=1.38.0",
    "langchain-text-splitters>=0.3.11",
    "loguru>=0.7.3",
    "mistralai>=1.9.10",
]

[tool.flake8]
max-line-length = 120
